renaming Lightning package, adding json imports to Visualization and moving lineStreaming function to Plots.
