+++ src/main/scala/org/viz/lightning/types/Plots.scala
@@ -44,6 +44,27 @@ trait Plots extends Base {
   }
 
   /**
+   * One or more one-dimensional series data streamed as lines.
+   */
+  def lineStreaming(series: Array[Array[Double]],
+                    size: Array[Double],
+                    color: Array[Array[Double]] = Array[Array[Double]](),
+                    alpha: Array[Double] = Array[Double] (),
+                    label: Array[Int] = Array[Int](),
+                    xaxis: String = "",
+                    yaxis: String = ""): Visualization = {
+
+    val data = Map("series" -> series.toList, "color" -> color.toList)
+
+    val settings = new Settings()
+      .append(List(Label(label), Size(size), Alpha(alpha)))
+      .append(List(Axis(xaxis, "xaxis"), Axis(yaxis, "yaxis")))
+
+    plot("line-streaming", data ++ settings.toMap)
+
+  }
+
+  /**
    * Force-directed network from connectivity.
    */
   def force(conn: Array[Array[Double]],
